Update jira.js to support any Jira project
The current jira.js implementation was hardcoded with lookups for
ER base, which was unecessary. There was also behavior where the
existing regex for issues without the ER project were being
skipped. This solves for both conditions, while still moving to
a search properly if a non-issue type is provided.